working router with usable dashboard,explorer,home routes
