+++ package.json
@@ -35,6 +35,7 @@
     "babel-eslint": "^7.2.3",
     "css-loader": "^0.28.1",
     "d3": "^4.8.0",
+    "history": "^4.6.1",
     "react": "^15.5.4",
     "react-desktop": "^0.3.0",
     "react-dnd": "^2.3.0",
@@ -45,6 +46,7 @@
     "react-resizable": "^1.7.0",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
+    "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
     "sequelize": "^3.30.4",
     "style-loader": "^0.17.0"
